requirements.txt
streamlit==1.52.2
pillow==12.0.0
numpy==2.3.5
torch==2.9.1
torchvision==0.24.1